Changed Rating table TotalScore type to float

+++ Storage/Tables/Rating.cs
@@ -7,7 +7,7 @@ namespace Storage.Tables
     public class Rating
     {
         public int Id { get; set; }
-        public int TotalScore { get; set; }
+        public float TotalScore { get; set; }
         [ForeignKey(nameof(Item))]
         public int ItemId { get; set; }
         public Item Item { get; set; }
